+++ site/plugins/indieConnector/lib/BlueskySender.php
@@ -22,7 +22,6 @@ class BlueskySender extends ExternalPostSender
 
     public function sendPost($page)
     {
-
         if (!$this->enabled) {
             return false;
         }
@@ -50,8 +49,8 @@ class BlueskySender extends ExternalPostSender
         }
 
         try {
-            $pageUrl = $page->url();
-            $trimTextPosition = $this->calculatePostTextLength($page->url());
+            $pageUrl = $this->getPostUrl($page);
+            $trimTextPosition = $this->calculatePostTextLength($pageUrl);
             $language = 'en';
 
             $message = $this->getTextFieldContent($page, $trimTextPosition);
@@ -59,6 +58,10 @@ class BlueskySender extends ExternalPostSender
 
             if ($defaultLanguage = kirby()->defaultLanguage()) {
                 $language = $defaultLanguage->code();
+            }
+
+            if ($this->prefereLanguage !== false && !empty($this->prefereLanguage)) {
+                $language = $this->prefereLanguage;
             }
 
             $bluesky = new BlueskyApi();
@@ -91,6 +94,10 @@ class BlueskySender extends ExternalPostSender
                         [
                             'alt' => $page->title()->value(),
                             'image' => $image,
+                            'aspectRatio' => [
+                                'width'  => $mediaAttachment['width'],
+                                'height' => $mediaAttachment['height'],
+                            ],
                         ],
                     ],
                 ];
@@ -101,8 +108,21 @@ class BlueskySender extends ExternalPostSender
 
             $response = $bluesky->request('POST', 'com.atproto.repo.createRecord', $args);
 
-            $this->updatePosts($response->uri, 200, $page, 'bluesky');
-            return true;
+            if (isset($response->error)) {
+                return [
+                    'id' => null,
+                    'uri' => null,
+                    'status' => 500,
+                    'target' => 'bluesky'
+                ];
+            }
+
+            return [
+                'id' => $response->cid,
+                'uri' => $response->uri,
+                'status' => 200,
+                'target' => 'bluesky'
+            ];
         } catch (Exception $e) {
             throw new Exception($e->getMessage());
             return false;
@@ -143,17 +163,21 @@ class BlueskySender extends ExternalPostSender
                 $imagefield = $this->imagefield;
                 $image = $page->$imagefield();
 
+
                 if (!is_null($image) && $image->isNotEmpty()) {
                     $imageMimeType = $image->toFile()->mime();
-                    $imagePath = $image->toFile()->resize(800)->root(); // TODO image size must be very low, so we need to resize it
+                    $resizedImage = $image->toFile()->resize(800); // image size must be very low, so we need to resize it
+                    $resizedImage->base64(); // this forces kirby to generate the image
 
-                    if (!F::exists($imagePath)) {
+                    if (!F::exists($resizedImage->root())) {
                         return false;
                     }
 
                     return [
-                        'content' => file_get_contents($imagePath),
+                        'content' => file_get_contents($resizedImage->root()),
                         'mime' => $imageMimeType,
+                        'width' => $resizedImage->width(),
+                        'height' => $resizedImage->height()
                     ];
                 }
             }