+++ site/plugins/indieConnector/lib/WebmentionSender.php
@@ -69,10 +69,10 @@ class WebmentionSender extends Sender
                 'status' => $status,
                 'retries' => 0,
             ];
-
         }
 
         $mergedUrls = $this->mergeUrlsWithOutbox($processedUrls, $page);
+
         $this->updateWebmentions($mergedUrls, $page);
 
         if (option('mauricerenck.indieConnector.stats.enabled', false)) {
@@ -82,6 +82,38 @@ class WebmentionSender extends Sender
 
         return $mergedUrls;
     }
+
+    public function sendWebmentionFromHook($page, $targetUrl, $sourceUrl)
+    {
+        // global config
+        if (!$this->activeWebmentions) {
+            return false;
+        }
+
+        $processedUrls = [];
+
+        $sent = $this->send($targetUrl, $sourceUrl);
+
+        $status = $sent ? 'success' : 'error';
+        $processedUrls[] = [
+            'url' => $targetUrl,
+            'date' => date('Y-m-d H:i:s'),
+            'status' => $status,
+            'retries' => 0,
+        ];
+
+        $mergedUrls = $this->mergeUrlsWithOutbox($processedUrls, $page);
+
+        $this->updateWebmentions($mergedUrls, $page);
+
+        if (option('mauricerenck.indieConnector.stats.enabled', false)) {
+            $stats = new WebmentionStats();
+            $stats->trackOutgoingWebmentions($mergedUrls, $page);
+        }
+
+        return $mergedUrls;
+    }
+
 
     public function send(string $targetUrl, string $sourceUrl)
     {
@@ -183,12 +215,12 @@ class WebmentionSender extends Sender
 
             $existingEntry =
                 count($existingEntries) === 0
-                    ? $newEntry
-                    : array_shift($existingEntries);
+                ? $newEntry
+                : array_shift($existingEntries);
 
             $mergedEntry = array_merge($existingEntry, $newEntry);
 
-            if($mergedEntry['status'] === 'error') {
+            if ($mergedEntry['status'] === 'error') {
                 $mergedEntry['retries'] = $existingEntry['retries'] + 1;
             }